Add do-until loop while electing Consul leader
The Vault configuration uses Consul as its storage backend. When
starting the Consul server agent, it must elect a leader; this process
usually takes a few seconds. If there is an attempt to initialize the
Vault BEFORE a leader is elected, then the server returns a 500
Internal Server Error response. This fix issues a request to the Get
Raft Leader endpoint until it returns the IP address of the leader in
the response and then proceeds to unseal the Vault.