Update Nomad Upstart configuration file
Nomad can integrate with Vault by defining a `vault` stanza in the
agent configuration file. The `vault` stanza requires the initial root
token for the integration to be successful. However, instead of writing
the initial root token to the agent configuration file, this commit
writes the initial root token to a file owned by the `vault` user and
`vault` group and is exported in the Nomad Upstart configuration file.

See https://bit.ly/2EVov9X for additional information.